# Backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BeeBa API
  version: 0.1.0
  description: Versioned API for the BeeBa Basis / VR content catalog.
servers:
  - url: http://localhost:8080
    description: Local backend
  - url: https://beeba.org
    description: Production domain
paths:
...
  /api/v1/categories:
    get:
      summary: List active public content categories
      operationId: listCategories
      responses:
        "200":
          description: Active categories ordered for catalog navigation.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/tags:
    get:
      summary: List public tags with published content counts
      operationId: listTags
      responses:
        "200":
          description: Public tags ordered by published content count.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tag"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/admin/users:
    get:
      summary: List users for admin management
      operationId: listAdminUsers
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
            maxLength: 200
        - name: role
          in: query
          schema:
            type: string
            enum:
              - user
              - moderator
              - admin
              - owner
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 50
      responses:
        "200":
          description: Active users visible to admin and owner roles.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminUser"
                  pagination:
                    $ref: "#/components/schemas/CursorPagination"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/admin/categories:
    get:
      summary: List categories for admin management
      operationId: listAdminCategories
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: All categories visible to admin and owner roles.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/admin/categories/{category_id}:
    patch:
      summary: Update category settings
      operationId: updateAdminCategory
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCategoryUpdateRequest"
      responses:
        "200":
          description: Category updated and audited.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/Category"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/admin/tags:
    get:
      summary: List tags for admin management
      operationId: listAdminTags
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Tags visible to admin and owner roles.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tag"
        default:
          $ref: "#/components/responses/Error"
    post:
      summary: Create a system or user-visible tag
      operationId: createAdminTag
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminTagCreateRequest"
      responses:
        "201":
          description: Tag created and audited.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/Tag"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/admin/tags/{tag_id}:
    patch:
      summary: Update tag settings
      operationId: updateAdminTag
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: tag_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminTagUpdateRequest"
      responses:
        "200":
          description: Tag updated and audited.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/Tag"
        default:
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: beeba_access_token
      description: Browser session cookie. Unsafe methods authenticated by cookie also require `X-CSRF-Token` matching the `beeba_csrf_token` cookie.
  responses:
    Error:
      description: Consistent API error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
    Category:
      type: object
      required:
        - id
        - slug
        - name
        - description
        - sort_order
        - is_active
        - published_count
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          enum:
            - worlds
            - avatars
            - props
            - prefabs
        name:
          type: string
        description:
          type: string
        icon_key:
          type:
            - string
            - "null"
        sort_order:
          type: integer
        is_active:
          type: boolean
        published_count:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Tag:
      type: object
      required:
        - id
        - slug
        - name
        - is_system
        - published_count
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        is_system:
          type: boolean
        published_count:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AdminCategoryUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        description:
          type: string
          maxLength: 500
        icon_key:
          type: string
          maxLength: 80
        sort_order:
          type: integer
          minimum: 0
          maximum: 10000
        is_active:
          type: boolean
    AdminTagCreateRequest:
      type: object
      required:
        - slug
        - name
      additionalProperties: false
      properties:
        slug:
          type: string
          maxLength: 80
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        name:
          type: string
          minLength: 2
          maxLength: 80
        is_system:
          type: boolean
          default: false
    AdminTagUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        is_system:
          type: boolean
...
    CursorPagination:
      type: object
      required:
        - limit
      properties:
        next_cursor:
          type:
            - string
            - "null"
        limit:
          type: integer
          minimum: 1
...
    AdminUser:
      type: object
      required:
        - id
        - email
        - username
        - roles
        - content_count
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        display_name:
          type:
            - string
            - "null"
        avatar_image_id:
          type:
            - string
            - "null"
          format: uuid
        email_verified_at:
          type:
            - string
            - "null"
          format: date-time
        roles:
          type: array
          items:
            type: string
            enum:
              - user
              - moderator
              - admin
              - owner
        banned_at:
          type:
            - string
            - "null"
          format: date-time
        deleted_at:
          type:
            - string
            - "null"
          format: date-time
        content_count:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...